+++ src/sat/smt/sat_smt.h
@@ -19,6 +19,7 @@ Author:
 #include "ast/ast_pp.h"
 #include "ast/ast_ll_pp.h"
 #include "sat/sat_solver.h"
+#include "sat/smt/sat_internalizer.h"
 
 namespace sat {
     
@@ -28,20 +29,6 @@ namespace sat {
         eframe(expr* e) : m_e(e), m_idx(0) {}
     };
 
-    class sat_internalizer {
-    public:
-        virtual ~sat_internalizer() {}
-        virtual bool is_bool_op(expr* e) const = 0;
-        virtual literal internalize(expr* e, bool learned) = 0;
-        virtual bool_var to_bool_var(expr* e) = 0;
-        virtual bool_var add_bool_var(expr* e)  = 0;
-        virtual void cache(app* t, literal l) = 0;
-        virtual void uncache(literal l) = 0;
-        virtual void push() = 0;
-        virtual void pop(unsigned n) = 0;
-        virtual void set_expr2var_replay(obj_map<expr, sat::bool_var>* r) = 0;
-    };
-    
     class constraint_base {
         extension* m_ex;
         unsigned   m_mem[0];
