+++ src/sat/tactic/CMakeLists.txt
@@ -1,6 +1,7 @@
 z3_add_component(sat_tactic
   SOURCES
     goal2sat.cpp
+    sat2goal.cpp
     sat_tactic.cpp
   COMPONENT_DEPENDENCIES
     sat
