+++ src/shell/dimacs_frontend.cpp
@@ -26,6 +26,7 @@ Revision History:
 #include "sat/sat_params.hpp"
 #include "sat/sat_solver.h"
 #include "sat/tactic/goal2sat.h"
+#include "sat/tactic/sat2goal.h"
 #include "ast/reg_decl_plugins.h"
 #include "tactic/tactic.h"
 #include "tactic/fd_solver/fd_solver.h"
